Make httpd serve docs from /var/www/html etc.
The default docs are still installed in /usr/share/httpd/*, but
on first launch, they are copied over as a template to /var/www.

This shares /var/www/html (not htdocs) together with nginx, where
the webroot-setup.service lives. That one will be adjusted to also
copy apache files over to the dataroot.